requirements.txt
argon2-cffi
flask
sqlalchemy